requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.1.0
python-dotenv>=1.0.0
streamlit>=1.32.0
PyPDF2>=3.0.0
pandas>=2.0.0